Move realized gains to its own handler
